+++ src/index.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Overview from './Overview';
+import Overview from './Overview/Overview.js';
 import 'materialize-css/dist/css/materialize.css';
 import 'materialize-css/dist/js/materialize.js';
 import './index.css';
-import SonarqubeConnector from './connectors/SonarqubeConnector.js';
-import RestInterface from './connectors/RestInterface.js';
+import SonarqubeConnector from './Connectors/SonarqubeConnector.js';
+import RestInterface from './Connectors/RestInterface.js';
 
 const sonarqubeClient = new SonarqubeConnector("http://localhost:8091/https://bpt-lab.org/sonarqube");
